# swagger-complete.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "🏢 University Housing System API",
    "version": "2.0.0",
    "description": "Complete API documentation for University Housing System with all endpoints",
    "contact": {
      "name": "Development Team",
      "email": "support@housing.edu.eg"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development"
    },
    {
      "url": "https://api.housing.edu.eg",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID Token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/applications": {
      "post": {
        "tags": ["Applications"],
        "summary": "Submit Housing Application",
        "description": "Submit a new housing application (Student only)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["fullName", "nationalId", "gender", "college", "academicYear"],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "nationalId": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string",
                    "enum": ["male", "female"]
                  },
                  "college": {
                    "type": "string"
                  },
                  "academicYear": {
                    "type": "number"
                  },
                  "gpa": {
                    "type": "number"
                  },
                  "nationalIdCard": {
                    "type": "string",
                    "format": "binary"
                  },
                  "personalPhoto": {
                    "type": "string",
                    "format": "binary"
                  },
                  "medicalReport": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully"
          },
          "400": {
            "description": "Validation error or active application already exists"
          }
        }
      },
      "get": {
        "tags": ["Applications"],
        "summary": "Get All Applications",
        "description": "Retrieve all applications (Admin/Supervisor only)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "rejected", "needs_update"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications retrieved"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get All Users",
        "description": "List all users with pagination and filtering (Admin only)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 20
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["student", "admin", "supervisor"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved"
          }
        }
      }
    },
...
    "/api/buildings": {
      "get": {
        "tags": ["Buildings"],
        "summary": "Get All Buildings",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Buildings retrieved"
          }
        }
      },
      "post": {
        "tags": ["Buildings"],
        "summary": "Create Building",
        "description": "Create a new building (Admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "gender"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string",
                    "enum": ["male", "female"]
                  },
                  "location": {
                    "type": "string"
                  },
                  "floorCount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Building created"
          }
        }
      }
    },
...
    "/api/rooms": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Get All Rooms",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "buildingId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rooms retrieved"
          }
        }
      },
      "post": {
        "tags": ["Rooms"],
        "summary": "Create Room",
        "description": "Create a new room (Admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roomNumber", "buildingId", "capacity"],
                "properties": {
                  "roomNumber": {
                    "type": "string"
                  },
                  "buildingId": {
                    "type": "string"
                  },
                  "floorNumber": {
                    "type": "number"
                  },
                  "capacity": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created"
          }
        }
      }
    },
...
    "/api/housing-requests": {
      "post": {
        "tags": ["Housing Requests"],
        "summary": "Submit Housing Request",
        "description": "Submit transfer, leave, vacate, or maintenance request (Student only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type"],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["transfer", "leave", "vacate", "maintenance"]
                  },
                  "toRoomId": {
                    "type": "string",
                    "description": "Required for transfer requests"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Required for leave/vacate requests"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Required for leave/vacate requests"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request submitted successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "tags": ["Housing Requests"],
        "summary": "Get All Requests",
        "description": "Retrieve all housing requests (Admin only)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requests retrieved"
          }
        }
      }
    },
...
    "/api/meals": {
      "get": {
        "tags": ["Meals"],
        "summary": "Get Meals",
        "description": "Retrieve meal menu for a specific date",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["breakfast", "lunch", "dinner"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meals retrieved"
          }
        }
      },
      "post": {
        "tags": ["Meals"],
        "summary": "Create Meal",
        "description": "Add new meal to menu (Admin/Meal Admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "type", "date"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["breakfast", "lunch", "dinner"]
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  },
                  "nutritionInfo": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meal created"
          }
        }
      }
    },
...
    "/api/reports": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get All Reports",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reports retrieved"
          }
        }
      },
      "post": {
        "tags": ["Reports"],
        "summary": "Submit Report",
        "description": "Submit a complaint or maintenance report (Student)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "title", "description"],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["complaint", "maintenance", "other"]
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["low", "medium", "high"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report submitted"
          }
        }
      }
    },
...
    "/api/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get All Payments",
        "description": "Get payments list (Admin only)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payments retrieved"
          }
        }
      },
      "post": {
        "tags": ["Payments"],
        "summary": "Create Payment",
        "description": "Create a new payment (Student only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "paymentMethod"],
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": ["cash", "card", "bank_transfer", "online"]
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created"
          }
        }
      }
    },
...
    "/api/announcements": {
      "get": {
        "tags": ["Announcements"],
        "summary": "Get Announcements",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Announcements retrieved"
          }
        }
      },
      "post": {
        "tags": ["Announcements"],
        "summary": "Create Announcement",
        "description": "Create announcement (Admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Announcement created"
          }
        }
      }
    },
...
  }
}